+++ frontend/src/layouts/HomeLayout/_defaultProps.tsx
@@ -1,4 +1,5 @@
 import {
+  AppstoreFilled,
   DatabaseFilled,
   GlobalOutlined,
   HomeFilled,
@@ -17,6 +18,11 @@ export default {
         path: '/home',
         nameKey: 'homeLayout.route.home',
         icon: <HomeFilled style={brandIcon()} />,
+      },
+      {
+        path: '/catalog',
+        nameKey: 'homeLayout.route.catalog',
+        icon: <AppstoreFilled style={brandIcon()} />,
       },
       {
         path: '/dataModels',